requirements.txt
langchain==0.0.340
chromadb==0.4.18
sentence_transformers
pypdf
python-box